feat: add @fuel-explorer/contract-ids
