+++ packages/contract-ids/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@fuel-explorer/contract-ids",
+  "private": true,
+  "version": "0.0.1",
+  "main": "./src/index.ts",
+  "module": "./src/index.ts",
+  "types": "./src/index.ts",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "scripts": {
+    "ts:check": "tsc --noEmit"
+  }
+}
