+++ packages/graphql-new/package.json
@@ -55,7 +55,7 @@
     "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "@wagmi/core": "2.6.10",
-    "app-commons": "workspace:*",
+    "@fuel-explorer/contract-ids": "workspace:*",
     "cors": "^2.8.5",
     "dayjs": "1.11.10",
     "dotenv": "16.4.4",
